+++ pandas/tests/tseries/offsets/test_custom_business_hour.py
@@ -1,0 +1,293 @@
+"""
+Tests for offsets.CustomBusinessHour
+"""
+from datetime import datetime
+
+import numpy as np
+import pytest
+
+from pandas._libs.tslibs import Timestamp
+from pandas._libs.tslibs.offsets import BusinessHour, CustomBusinessHour, Nano
+
+import pandas._testing as tm
+from pandas.tests.tseries.offsets.common import Base, assert_offset_equal
+
+
+class TestCustomBusinessHour(Base):
+    _offset = CustomBusinessHour
+    holidays = ["2014-06-27", datetime(2014, 6, 30), np.datetime64("2014-07-02")]
+
+    def setup_method(self, method):
+        # 2014 Calendar to check custom holidays
+        #   Sun Mon Tue Wed Thu Fri Sat
+        #  6/22  23  24  25  26  27  28
+        #    29  30 7/1   2   3   4   5
+        #     6   7   8   9  10  11  12
+        self.d = datetime(2014, 7, 1, 10, 00)
+        self.offset1 = CustomBusinessHour(weekmask="Tue Wed Thu Fri")
+
+        self.offset2 = CustomBusinessHour(holidays=self.holidays)
+
+    def test_constructor_errors(self):
+        from datetime import time as dt_time
+
+        msg = "time data must be specified only with hour and minute"
+        with pytest.raises(ValueError, match=msg):
+            CustomBusinessHour(start=dt_time(11, 0, 5))
+        msg = "time data must match '%H:%M' format"
+        with pytest.raises(ValueError, match=msg):
+            CustomBusinessHour(start="AAA")
+        msg = "time data must match '%H:%M' format"
+        with pytest.raises(ValueError, match=msg):
+            CustomBusinessHour(start="14:00:05")
+
+    def test_different_normalize_equals(self):
+        # GH#21404 changed __eq__ to return False when `normalize` does not match
+        offset = self._offset()
+        offset2 = self._offset(normalize=True)
+        assert offset != offset2
+
+    def test_repr(self):
+        assert repr(self.offset1) == "<CustomBusinessHour: CBH=09:00-17:00>"
+        assert repr(self.offset2) == "<CustomBusinessHour: CBH=09:00-17:00>"
+
+    def test_with_offset(self):
+        expected = Timestamp("2014-07-01 13:00")
+
+        assert self.d + CustomBusinessHour() * 3 == expected
+        assert self.d + CustomBusinessHour(n=3) == expected
+
+    def test_eq(self):
+        for offset in [self.offset1, self.offset2]:
+            assert offset == offset
+
+        assert CustomBusinessHour() != CustomBusinessHour(-1)
+        assert CustomBusinessHour(start="09:00") == CustomBusinessHour()
+        assert CustomBusinessHour(start="09:00") != CustomBusinessHour(start="09:01")
+        assert CustomBusinessHour(start="09:00", end="17:00") != CustomBusinessHour(
+            start="17:00", end="09:01"
+        )
+
+        assert CustomBusinessHour(weekmask="Tue Wed Thu Fri") != CustomBusinessHour(
+            weekmask="Mon Tue Wed Thu Fri"
+        )
+        assert CustomBusinessHour(holidays=["2014-06-27"]) != CustomBusinessHour(
+            holidays=["2014-06-28"]
+        )
+
+    def test_sub(self):
+        # override the Base.test_sub implementation because self.offset2 is
+        # defined differently in this class than the test expects
+        pass
+
+    def test_hash(self):
+        assert hash(self.offset1) == hash(self.offset1)
+        assert hash(self.offset2) == hash(self.offset2)
+
+    def test_call(self):
+        with tm.assert_produces_warning(FutureWarning):
+            # GH#34171 DateOffset.__call__ is deprecated
+            assert self.offset1(self.d) == datetime(2014, 7, 1, 11)
+            assert self.offset2(self.d) == datetime(2014, 7, 1, 11)
+
+    def testRollback1(self):
+        assert self.offset1.rollback(self.d) == self.d
+        assert self.offset2.rollback(self.d) == self.d
+
+        d = datetime(2014, 7, 1, 0)
+
+        # 2014/07/01 is Tuesday, 06/30 is Monday(holiday)
+        assert self.offset1.rollback(d) == datetime(2014, 6, 27, 17)
+
+        # 2014/6/30 and 2014/6/27 are holidays
+        assert self.offset2.rollback(d) == datetime(2014, 6, 26, 17)
+
+    def testRollback2(self):
+        assert self._offset(-3).rollback(datetime(2014, 7, 5, 15, 0)) == datetime(
+            2014, 7, 4, 17, 0
+        )
+
+    def testRollforward1(self):
+        assert self.offset1.rollforward(self.d) == self.d
+        assert self.offset2.rollforward(self.d) == self.d
+
+        d = datetime(2014, 7, 1, 0)
+        assert self.offset1.rollforward(d) == datetime(2014, 7, 1, 9)
+        assert self.offset2.rollforward(d) == datetime(2014, 7, 1, 9)
+
+    def testRollforward2(self):
+        assert self._offset(-3).rollforward(datetime(2014, 7, 5, 16, 0)) == datetime(
+            2014, 7, 7, 9
+        )
+
+    def test_roll_date_object(self):
+        offset = BusinessHour()
+
+        dt = datetime(2014, 7, 6, 15, 0)
+
+        result = offset.rollback(dt)
+        assert result == datetime(2014, 7, 4, 17)
+
+        result = offset.rollforward(dt)
+        assert result == datetime(2014, 7, 7, 9)
+
+    normalize_cases = [
+        (
+            CustomBusinessHour(normalize=True, holidays=holidays),
+            {
+                datetime(2014, 7, 1, 8): datetime(2014, 7, 1),
+                datetime(2014, 7, 1, 17): datetime(2014, 7, 3),
+                datetime(2014, 7, 1, 16): datetime(2014, 7, 3),
+                datetime(2014, 7, 1, 23): datetime(2014, 7, 3),
+                datetime(2014, 7, 1, 0): datetime(2014, 7, 1),
+                datetime(2014, 7, 4, 15): datetime(2014, 7, 4),
+                datetime(2014, 7, 4, 15, 59): datetime(2014, 7, 4),
+                datetime(2014, 7, 4, 16, 30): datetime(2014, 7, 7),
+                datetime(2014, 7, 5, 23): datetime(2014, 7, 7),
+                datetime(2014, 7, 6, 10): datetime(2014, 7, 7),
+            },
+        ),
+        (
+            CustomBusinessHour(-1, normalize=True, holidays=holidays),
+            {
+                datetime(2014, 7, 1, 8): datetime(2014, 6, 26),
+                datetime(2014, 7, 1, 17): datetime(2014, 7, 1),
+                datetime(2014, 7, 1, 16): datetime(2014, 7, 1),
+                datetime(2014, 7, 1, 10): datetime(2014, 6, 26),
+                datetime(2014, 7, 1, 0): datetime(2014, 6, 26),
+                datetime(2014, 7, 7, 10): datetime(2014, 7, 4),
+                datetime(2014, 7, 7, 10, 1): datetime(2014, 7, 7),
+                datetime(2014, 7, 5, 23): datetime(2014, 7, 4),
+                datetime(2014, 7, 6, 10): datetime(2014, 7, 4),
+            },
+        ),
+        (
+            CustomBusinessHour(
+                1, normalize=True, start="17:00", end="04:00", holidays=holidays
+            ),
+            {
+                datetime(2014, 7, 1, 8): datetime(2014, 7, 1),
+                datetime(2014, 7, 1, 17): datetime(2014, 7, 1),
+                datetime(2014, 7, 1, 23): datetime(2014, 7, 2),
+                datetime(2014, 7, 2, 2): datetime(2014, 7, 2),
+                datetime(2014, 7, 2, 3): datetime(2014, 7, 3),
+                datetime(2014, 7, 4, 23): datetime(2014, 7, 5),
+                datetime(2014, 7, 5, 2): datetime(2014, 7, 5),
+                datetime(2014, 7, 7, 2): datetime(2014, 7, 7),
+                datetime(2014, 7, 7, 17): datetime(2014, 7, 7),
+            },
+        ),
+    ]
+
+    @pytest.mark.parametrize("norm_cases", normalize_cases)
+    def test_normalize(self, norm_cases):
+        offset, cases = norm_cases
+        for dt, expected in cases.items():
+            assert offset.apply(dt) == expected
+
+    def test_is_on_offset(self):
+        tests = [
+            (
+                CustomBusinessHour(start="10:00", end="15:00", holidays=self.holidays),
+                {
+                    datetime(2014, 7, 1, 9): False,
+                    datetime(2014, 7, 1, 10): True,
+                    datetime(2014, 7, 1, 15): True,
+                    datetime(2014, 7, 1, 15, 1): False,
+                    datetime(2014, 7, 5, 12): False,
+                    datetime(2014, 7, 6, 12): False,
+                },
+            )
+        ]
+
+        for offset, cases in tests:
+            for dt, expected in cases.items():
+                assert offset.is_on_offset(dt) == expected
+
+    apply_cases = [
+        (
+            CustomBusinessHour(holidays=holidays),
+            {
+                datetime(2014, 7, 1, 11): datetime(2014, 7, 1, 12),
+                datetime(2014, 7, 1, 13): datetime(2014, 7, 1, 14),
+                datetime(2014, 7, 1, 15): datetime(2014, 7, 1, 16),
+                datetime(2014, 7, 1, 19): datetime(2014, 7, 3, 10),
+                datetime(2014, 7, 1, 16): datetime(2014, 7, 3, 9),
+                datetime(2014, 7, 1, 16, 30, 15): datetime(2014, 7, 3, 9, 30, 15),
+                datetime(2014, 7, 1, 17): datetime(2014, 7, 3, 10),
+                datetime(2014, 7, 2, 11): datetime(2014, 7, 3, 10),
+                # out of business hours
+                datetime(2014, 7, 2, 8): datetime(2014, 7, 3, 10),
+                datetime(2014, 7, 2, 19): datetime(2014, 7, 3, 10),
+                datetime(2014, 7, 2, 23): datetime(2014, 7, 3, 10),
+                datetime(2014, 7, 3, 0): datetime(2014, 7, 3, 10),
+                # saturday
+                datetime(2014, 7, 5, 15): datetime(2014, 7, 7, 10),
+                datetime(2014, 7, 4, 17): datetime(2014, 7, 7, 10),
+                datetime(2014, 7, 4, 16, 30): datetime(2014, 7, 7, 9, 30),
+                datetime(2014, 7, 4, 16, 30, 30): datetime(2014, 7, 7, 9, 30, 30),
+            },
+        ),
+        (
+            CustomBusinessHour(4, holidays=holidays),
+            {
+                datetime(2014, 7, 1, 11): datetime(2014, 7, 1, 15),
+                datetime(2014, 7, 1, 13): datetime(2014, 7, 3, 9),
+                datetime(2014, 7, 1, 15): datetime(2014, 7, 3, 11),
+                datetime(2014, 7, 1, 16): datetime(2014, 7, 3, 12),
+                datetime(2014, 7, 1, 17): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 2, 11): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 2, 8): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 2, 19): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 2, 23): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 3, 0): datetime(2014, 7, 3, 13),
+                datetime(2014, 7, 5, 15): datetime(2014, 7, 7, 13),
+                datetime(2014, 7, 4, 17): datetime(2014, 7, 7, 13),
+                datetime(2014, 7, 4, 16, 30): datetime(2014, 7, 7, 12, 30),
+                datetime(2014, 7, 4, 16, 30, 30): datetime(2014, 7, 7, 12, 30, 30),
+            },
+        ),
+    ]
+
+    @pytest.mark.parametrize("apply_case", apply_cases)
+    def test_apply(self, apply_case):
+        offset, cases = apply_case
+        for base, expected in cases.items():
+            assert_offset_equal(offset, base, expected)
+
+    nano_cases = [
+        (
+            CustomBusinessHour(holidays=holidays),
+            {
+                Timestamp("2014-07-01 15:00")
+                + Nano(5): Timestamp("2014-07-01 16:00")
+                + Nano(5),
+                Timestamp("2014-07-01 16:00")
+                + Nano(5): Timestamp("2014-07-03 09:00")
+                + Nano(5),
+                Timestamp("2014-07-01 16:00")
+                - Nano(5): Timestamp("2014-07-01 17:00")
+                - Nano(5),
+            },
+        ),
+        (
+            CustomBusinessHour(-1, holidays=holidays),
+            {
+                Timestamp("2014-07-01 15:00")
+                + Nano(5): Timestamp("2014-07-01 14:00")
+                + Nano(5),
+                Timestamp("2014-07-01 10:00")
+                + Nano(5): Timestamp("2014-07-01 09:00")
+                + Nano(5),
+                Timestamp("2014-07-01 10:00")
+                - Nano(5): Timestamp("2014-06-26 17:00")
+                - Nano(5),
+            },
+        ),
+    ]
+
+    @pytest.mark.parametrize("nano_case", nano_cases)
+    def test_apply_nanoseconds(self, nano_case):
+        offset, cases = nano_case
+        for base, expected in cases.items():
+            assert_offset_equal(offset, base, expected)
