+++ pandas/tests/tseries/offsets/test_yqm_offsets.py
@@ -7,6 +7,11 @@ import pytest
 
 import pandas as pd
 from pandas import Timestamp
+from pandas.tests.tseries.offsets.common import (
+    Base,
+    assert_is_on_offset,
+    assert_offset_equal,
+)
 
 from pandas.tseries.offsets import (
     BMonthBegin,
@@ -22,9 +27,6 @@ from pandas.tseries.offsets import (
     YearBegin,
     YearEnd,
 )
-
-from .common import assert_is_on_offset, assert_offset_equal
-from .test_offsets import Base
 
 # --------------------------------------------------------------------
 # Misc
